+++ src/main/java/radon/jujutsu_kaisen/entity/base/SorcererEntity.java
@@ -13,6 +13,14 @@ import net.minecraft.world.entity.MobSpawnType;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
+import net.minecraft.world.entity.ai.goal.RandomLookAroundGoal;
+import net.minecraft.world.entity.ai.goal.target.HurtByTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.OwnerHurtByTargetGoal;
+import net.minecraft.world.entity.ai.goal.target.OwnerHurtTargetGoal;
+import net.minecraft.world.entity.animal.IronGolem;
+import net.minecraft.world.entity.npc.AbstractVillager;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.levelgen.structure.Structure;
@@ -22,6 +30,7 @@ import org.jetbrains.annotations.NotNull;
 import radon.jujutsu_kaisen.capability.data.ISorcererData;
 import radon.jujutsu_kaisen.capability.data.SorcererDataHandler;
 import radon.jujutsu_kaisen.capability.data.sorcerer.SorcererGrade;
+import radon.jujutsu_kaisen.entity.ai.goal.*;
 import radon.jujutsu_kaisen.util.HelperMethods;
 import software.bernie.geckolib.animatable.GeoEntity;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
@@ -47,7 +56,39 @@ public abstract class SorcererEntity extends PathfinderMob implements GeoEntity,
         return true;
     }
 
+    protected abstract boolean isCustom();
 
+    protected boolean canFly() { return false; }
+
+    protected boolean hasMeleeAttack() { return true; }
+
+    protected boolean targetsCurses() { return true; }
+    protected boolean targetsSorcerers() { return false; }
+
+    private void createGoals() {
+        int target = 1;
+        int goal = 1;
+
+        this.goalSelector.addGoal(goal++, new BetterFloatGoal(this));
+
+        if (this.hasMeleeAttack()) {
+            this.goalSelector.addGoal(goal++, new MeleeAttackGoal(this, 1.2D, true));
+        }
+        this.goalSelector.addGoal(goal++, new LookAtTargetGoal(this));
+        this.goalSelector.addGoal(goal++, this.canPerformSorcery() || !this.getCustom().isEmpty() ? new SorcererGoal(this) : new HealingGoal(this));
+        this.goalSelector.addGoal(goal, new RandomLookAroundGoal(this));
+
+        this.targetSelector.addGoal(target++, new HurtByTargetGoal(this));
+        this.targetSelector.addGoal(target++, new NearestAttackableTargetGoal<>(this, IronGolem.class, false));
+        this.targetSelector.addGoal(target++, new NearestAttackableTargetGoal<>(this, AbstractVillager.class, false));
+
+        if (this.targetsSorcerers()) {
+            this.targetSelector.addGoal(target++, new NearestAttackableSorcererGoal(this, true));
+        }
+        if (this.targetsCurses()) {
+            this.targetSelector.addGoal(target, new NearestAttackableCurseGoal(this, true));
+        }
+    }
 
     @Override
     public boolean isPersistenceRequired() {
@@ -116,6 +157,13 @@ public abstract class SorcererEntity extends PathfinderMob implements GeoEntity,
         super.aiStep();
     }
 
+    @Override
+    protected void customServerAiStep() {
+        super.customServerAiStep();
+
+        this.setSprinting(this.getDeltaMovement().lengthSqr() > 0.01D && this.moveControl.getSpeedModifier() > 1.0D);
+    }
+
     public static AttributeSupplier.Builder createAttributes() {
         return SorcererEntity.createMobAttributes()
                 .add(Attributes.MOVEMENT_SPEED, 0.32D)
@@ -126,6 +174,8 @@ public abstract class SorcererEntity extends PathfinderMob implements GeoEntity,
     @Override
     public void onAddedToWorld() {
         super.onAddedToWorld();
+
+        if (!this.isCustom()) this.createGoals();
 
         this.getCapability(SorcererDataHandler.INSTANCE).ifPresent(this::init);
     }