+++ src/main/java/radon/jujutsu_kaisen/entity/curse/DagonEntity.java
@@ -19,6 +19,7 @@ import java.util.List;
 
 public class DagonEntity extends DisasterCurse {
     private static final RawAnimation WALK = RawAnimation.begin().thenLoop("move.walk");
+    private static final RawAnimation RUN = RawAnimation.begin().thenLoop("move.run");
     private static final RawAnimation SWING = RawAnimation.begin().thenPlay("attack.swing");
 
     public DagonEntity(EntityType<? extends TamableAnimal> pEntityType, Level pLevel) {
@@ -31,17 +32,7 @@ public class DagonEntity extends DisasterCurse {
     }
 
     @Override
-    protected boolean canFly() {
-        return false;
-    }
-
-    @Override
     public boolean canPerformSorcery() {
-        return true;
-    }
-
-    @Override
-    protected boolean hasMeleeAttack() {
         return true;
     }
 
@@ -65,9 +56,9 @@ public class DagonEntity extends DisasterCurse {
         return JJKAbilities.HORIZON_OF_THE_CAPTIVATING_SKANDHA.get();
     }
 
-    private PlayState walkPredicate(AnimationState<DagonEntity> animationState) {
+    private PlayState walkRunPredicate(AnimationState<DagonEntity> animationState) {
         if (animationState.isMoving()) {
-            return animationState.setAndContinue(WALK);
+            return animationState.setAndContinue(this.isSprinting() ? RUN : WALK);
         }
         return PlayState.STOP;
     }
@@ -82,7 +73,7 @@ public class DagonEntity extends DisasterCurse {
 
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllerRegistrar) {
-        controllerRegistrar.add(new AnimationController<>(this, "Walk", this::walkPredicate));
+        controllerRegistrar.add(new AnimationController<>(this, "Walk/Run", this::walkRunPredicate));
         controllerRegistrar.add(new AnimationController<>(this, "Swing", this::swingPredicate));
     }
 }