+++ src/main/java/radon/jujutsu_kaisen/entity/curse/FelineCurseEntity.java
@@ -103,22 +103,12 @@ public class FelineCurseEntity extends PackCursedSpirit {
 
     @Override
     protected boolean isCustom() {
-        return true;
-    }
-
-    @Override
-    protected boolean canFly() {
         return false;
     }
 
     @Override
     public boolean canPerformSorcery() {
         return false;
-    }
-
-    @Override
-    protected boolean hasMeleeAttack() {
-        return true;
     }
 
     @Override