+++ src/main/java/radon/jujutsu_kaisen/entity/curse/ZombaCurseEntity.java
@@ -32,18 +32,8 @@ public class ZombaCurseEntity extends CursedSpirit {
     }
 
     @Override
-    protected boolean canFly() {
-        return false;
-    }
-
-    @Override
     public boolean canPerformSorcery() {
         return false;
-    }
-
-    @Override
-    protected boolean hasMeleeAttack() {
-        return true;
     }
 
     @Override